# Fitness-ai-frontend/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ExerciseRecordCreate": {
        "properties": {
          "count": {
            "description": "完成次数",
            "minimum": 0.0,
            "title": "Count",
            "type": "integer"
          },
          "duration": {
            "description": "时长 (秒)",
            "minimum": 0.0,
            "title": "Duration",
            "type": "integer"
          },
          "exercise_id": {
            "title": "Exercise Id",
            "type": "integer"
          },
          "feedback": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feedback"
          },
          "heart_rate_avg": {
            "anyOf": [
              {
                "maximum": 260.0,
                "minimum": 20.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Heart Rate Avg"
          },
          "heart_rate_max": {
            "anyOf": [
              {
                "maximum": 260.0,
                "minimum": 20.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Heart Rate Max"
          },
          "keypoints_data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Keypoints Data"
          },
          "score": {
            "description": "动作评分 0-100",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Score",
            "type": "number"
          }
        },
        "required": [
          "exercise_id",
          "score",
          "count",
          "duration"
        ],
        "title": "ExerciseRecordCreate",
        "type": "object"
      },
      "ExerciseRecordPage": {
        "description": "分页运动记录响应。",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ExerciseRecordResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "skip": {
            "title": "Skip",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "skip",
          "limit"
        ],
        "title": "ExerciseRecordPage",
        "type": "object"
      },
      "ExerciseRecordResponse": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "duration": {
            "title": "Duration",
            "type": "integer"
          },
          "exercise_id": {
            "title": "Exercise Id",
            "type": "integer"
          },
          "feedback": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feedback"
          },
          "heart_rate_avg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Heart Rate Avg"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "score": {
            "title": "Score",
            "type": "number"
          },
          "video_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Video Url"
          }
        },
        "required": [
          "id",
          "exercise_id",
          "score",
          "count",
          "duration",
          "heart_rate_avg",
          "video_url",
          "feedback",
          "created_at"
        ],
        "title": "ExerciseRecordResponse",
        "type": "object"
      },
      "ExerciseRecordUpdate": {
        "properties": {
          "count": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "完成次数",
            "title": "Count"
          },
          "duration": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "时长 (秒)",
            "title": "Duration"
          },
          "feedback": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feedback"
          },
          "heart_rate_avg": {
            "anyOf": [
              {
                "maximum": 260.0,
                "minimum": 20.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Heart Rate Avg"
          },
          "heart_rate_max": {
            "anyOf": [
              {
                "maximum": 260.0,
                "minimum": 20.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Heart Rate Max"
          },
          "keypoints_data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Keypoints Data"
          },
          "score": {
            "anyOf": [
              {
                "maximum": 100.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "动作评分 0-100",
            "title": "Score"
          }
        },
        "title": "ExerciseRecordUpdate",
        "type": "object"
      },
      "ExerciseResponse": {
        "properties": {
          "aliases": {
            "items": {
              "type": "string"
            },
            "title": "Aliases",
            "type": "array"
          },
          "analysis_rule_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analysis Rule Version"
          },
          "analysis_status_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analysis Status Reason"
          },
          "analysis_supported": {
            "default": false,
            "title": "Analysis Supported",
            "type": "boolean"
          },
          "body_part": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Body Part"
          },
          "campus_candidate_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Campus Candidate Reason"
          },
          "canonical_action_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Canonical Action Key"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "equipment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Equipment"
          },
          "external_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Id"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "instruction_steps": {
            "additionalProperties": true,
            "title": "Instruction Steps",
            "type": "object"
          },
          "instructions": {
            "additionalProperties": true,
            "title": "Instructions",
            "type": "object"
          },
          "is_bodyweight": {
            "default": false,
            "title": "Is Bodyweight",
            "type": "boolean"
          },
          "is_low_equipment_candidate": {
            "default": false,
            "title": "Is Low Equipment Candidate",
            "type": "boolean"
          },
          "media_attribution": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Media Attribution"
          },
          "muscle_group": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Muscle Group"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "secondary_muscles": {
            "items": {
              "type": "string"
            },
            "title": "Secondary Muscles",
            "type": "array"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "target": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target"
          },
          "target_muscles": {
            "items": {
              "type": "string"
            },
            "title": "Target Muscles",
            "type": "array"
          }
        },
        "required": [
          "id",
          "name",
          "category",
          "description"
        ],
        "title": "ExerciseResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/auth/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "description": "校园健康体适能检测与管理系统",
    "title": "体适能 AI 管家 API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/exercise/exercises": {
      "get": {
        "description": "获取标准动作列表",
        "operationId": "get_exercises_api_exercise_exercises_get",
        "parameters": [
          {
            "description": "按名称、别名、部位、器械或肌群搜索",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "按名称、别名、部位、器械或肌群搜索",
              "title": "Q"
            }
          },
          {
            "description": "按外部动作目录 equipment 过滤",
            "in": "query",
            "name": "equipment",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "按外部动作目录 equipment 过滤",
              "title": "Equipment"
            }
          },
          {
            "description": "按外部动作目录 body_part 过滤",
            "in": "query",
            "name": "body_part",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "按外部动作目录 body_part 过滤",
              "title": "Body Part"
            }
          },
          {
            "description": "仅返回已接入 AI 评分规则的动作",
            "in": "query",
            "name": "analysis_supported",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "仅返回已接入 AI 评分规则的动作",
              "title": "Analysis Supported"
            }
          },
          {
            "description": "仅返回校园低器械候选动作",
            "in": "query",
            "name": "campus_candidate",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "仅返回校园低器械候选动作",
              "title": "Campus Candidate"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ExerciseResponse"
                  },
                  "title": "Response Get Exercises Api Exercise Exercises Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Exercises",
        "tags": [
          "运动"
        ]
      }
    },
    "/api/exercise/records": {
      "delete": {
        "description": "批量删除运动记录",
        "operationId": "batch_delete_records_api_exercise_records_delete",
        "parameters": [
          {
            "description": "要删除的记录 ID 列表",
            "in": "query",
            "name": "record_ids",
            "required": true,
            "schema": {
              "description": "要删除的记录 ID 列表",
              "items": {
                "type": "integer"
              },
              "title": "Record Ids",
              "type": "array"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Batch Delete Records",
        "tags": [
          "运动"
        ]
      },
      "get": {
        "description": "获取用户运动记录（支持日期范围、动作 ID 过滤）。",
        "operationId": "get_user_records_api_exercise_records_get",
        "parameters": [
          {
            "in": "query",
            "name": "start_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start Date"
            }
          },
          {
            "in": "query",
            "name": "end_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End Date"
            }
          },
          {
            "in": "query",
            "name": "exercise_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Exercise Id"
            }
          },
          {
            "description": "跳过的记录数",
            "in": "query",
            "name": "skip",
            "required": false,
            "schema": {
              "default": 0,
              "description": "跳过的记录数",
              "minimum": 0,
              "title": "Skip",
              "type": "integer"
            }
          },
          {
            "description": "返回的记录数，范围 1-100",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "description": "返回的记录数，范围 1-100",
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ExerciseRecordResponse"
                  },
                  "title": "Response Get User Records Api Exercise Records Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get User Records",
        "tags": [
          "运动"
        ]
      },
      "post": {
        "description": "创建运动记录",
        "operationId": "create_record_api_exercise_records_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExerciseRecordCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseRecordResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Create Record",
        "tags": [
          "运动"
        ]
      }
    },
    "/api/exercise/records/page": {
      "get": {
        "description": "获取用户运动记录分页响应（含总数和分页元数据）。",
        "operationId": "get_user_records_page_api_exercise_records_page_get",
        "parameters": [
          {
            "in": "query",
            "name": "start_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start Date"
            }
          },
          {
            "in": "query",
            "name": "end_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End Date"
            }
          },
          {
            "in": "query",
            "name": "exercise_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Exercise Id"
            }
          },
          {
            "description": "跳过的记录数",
            "in": "query",
            "name": "skip",
            "required": false,
            "schema": {
              "default": 0,
              "description": "跳过的记录数",
              "minimum": 0,
              "title": "Skip",
              "type": "integer"
            }
          },
          {
            "description": "返回的记录数，范围 1-100",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "description": "返回的记录数，范围 1-100",
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseRecordPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get User Records Page",
        "tags": [
          "运动"
        ]
      }
    },
    "/api/exercise/records/{record_id}": {
      "delete": {
        "description": "删除运动记录",
        "operationId": "delete_record_api_exercise_records__record_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "record_id",
            "required": true,
            "schema": {
              "title": "Record Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Delete Record",
        "tags": [
          "运动"
        ]
      },
      "get": {
        "description": "获取单条运动记录详情",
        "operationId": "get_record_detail_api_exercise_records__record_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "record_id",
            "required": true,
            "schema": {
              "title": "Record Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseRecordResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get Record Detail",
        "tags": [
          "运动"
        ]
      },
      "put": {
        "description": "修改运动记录",
        "operationId": "update_record_api_exercise_records__record_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "record_id",
            "required": true,
            "schema": {
              "title": "Record Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExerciseRecordUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseRecordResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Update Record",
        "tags": [
          "运动"
        ]
      }
    },
...
  }
}